refactor: Phase 3 - Extract UI controls panel into focused modules
- Refactor grouped-simulation-controls-panel.tsx from 678 lines to 118 lines (82.6% reduction)
- Extract control grouping logic into pattern-control-grouping.ts (231 lines)
- Extract base control rendering into base-control-renderer.tsx (102 lines)
- Extract pattern-specific controls into pattern-specific-controls.tsx (175 lines)
- Maintain zero breaking changes to public APIs
- Preserve all pattern-specific behaviors (four-pole gradient, cellular automaton)
- All existing functionality and tests preserved
- Complete 4-command preflight validation passed

Module breakdown:
- Main panel: 118 lines (orchestrates rendering)
- Control grouping: 231 lines (pattern-specific grouping logic)
- Base renderer: 102 lines (standard control types)
- Pattern-specific: 175 lines (specialized control components)
